Phase A: extract output/write_occup_m as free functions in dftu_io namespace
- New files: source_pw/module_pwdft/dftu_output.h, dftu_output.cpp
  - declare and implement dftu_io::output and dftu_io::write_occup_m
  - first parameter is const Plus_U_Base&, accesses state via public getters
  - JacobiRotate and CalculateEigenvalues inline helpers migrated here
- dftu_base.h: added public getters get_U_Yukawa / get_J_Yukawa;
  removed output/write_occup_m member declarations
- dftu_base.cpp: removed output/write_occup_m implementations and the
  two Jacobi inline helpers (no longer referenced here)
- Callers updated:
  - setup_dftu_pw.cpp: dftu.output(...) -> dftu_io::output(dftu, ...)
  - setup_dftu_lcao.cpp: dftu_ptr->output(...) -> dftu_io::output(*dftu_ptr, ...)
- CMakeLists.txt and Makefile.Objects: added dftu_output.cpp/dftu_output.o

All 4 dftu unit tests pass.
