+++ source/Makefile.Objects
@@ -742,6 +742,7 @@ OBJS_SRCPW=h_ewald_pw.o\
     setup_pwwfc.o\
     update_cell_pw.o\
     dftu_base.o\
+    dftu_output.o\
     dftu_pw.o\
     setup_dftu_pw.o\
     deltaspin_pw.o\
