+++ source/source_pw/module_pwdft/setup_dftu_pw.cpp
@@ -1,5 +1,6 @@
 #include "source_pw/module_pwdft/setup_dftu_pw.h"
 #include "source_pw/module_pwdft/dftu_base.h" // mohan add 2025-11-06
+#include "source_pw/module_pwdft/dftu_output.h" // mohan add 2025-11-08
 #include "source_io/module_parameter/parameter.h"
 
 namespace pw
@@ -28,7 +29,7 @@ void iter_init_dftu_pw(const int iter,
     {
         dftu.cal_occ_pw(iter, psi, wg, ucell, p_chgmix, isk);
     }
-    dftu.output(ucell, PARAM.inp.out_chg[0], PARAM.globalv.global_out_dir, PARAM.inp.nspin, PARAM.globalv.npol);
+    dftu_io::output(dftu, ucell, PARAM.inp.out_chg[0], PARAM.globalv.global_out_dir, PARAM.inp.nspin, PARAM.globalv.npol);
 }
 
 }
